requirements.txt
numpy>=1.24,<2
scipy>=1.6,<2
pandas>=2.0
matplotlib>=3.0
tqdm>=4.0
preflibtools>=2.0.0
requests>=2.28.0
scikit-surprise>=1.1.0
